rename hw_1.py to hw1.py
