+++ hw1.py
