+++ test_hw1.py
@@ -1,9 +1,9 @@
+"""Tests for hw1.py"""
 import pytest
-from hw_1 import SalaryStats, get_salary_stats
+from hw1 import SalaryStats, get_salary_stats
 
-# (args, kwargs, expected)
 test_data = (
-    (tuple(), {}, SalaryStats(maximum=0.0, average=0.0, median=0.0)),
+    (tuple(), {}, SalaryStats(maximum=0, average=0, median=0)),
 )
